requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
attrs==21.2.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
defusedxml==0.7.1
Flask==2.0.1
Flask-RESTful==0.3.9
gevent==21.1.2
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
osm2geojson==0.1.29
psycopg2-binary==2.8.6
pycparser==2.20
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
Shapely==1.7.1
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
zeep==3.4.0
zope.event==4.5.0
zope.interface==5.4.0
python-dotenv==0.19.0
